pyproject.toml
[project]
name = "birdclef2026"
version = "0.1.0"
description = "BirdCLEF 2026 - Acoustic Species Identification in the Pantanal"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "soundfile>=0.12",
    "torch>=2.0",
    "scikit-learn>=1.3",
    "pyyaml>=6.0",
    "tqdm>=4.65",
    "kagglehub>=1.0.0",
    "tensorflow>=2.21.0",
    "pyarrow>=23.0.1",
]

[project.optional-dependencies]
tf = ["tensorflow>=2.15"]
neuropt = ["neuropt", "anthropic"]
dev = [
    "pytest>=7.0",
    "mkdocs-material>=9.0",
    "nbformat>=5.9",
    "pyflakes>=3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "mkdocs-material>=9.7.6",
    "pytest>=9.0.2",
]
